refactor : create LottoStore
extract purchase from LottoGenerator
LottoStore contains LottoGenerator to use creating List<Lotto>
